TS/TJ: DeleteIdsAction (clear) will also honour the closing hours.
git-svn-id: file:///home/thijs/development/meresco2git/meresco/meresco-harvester/trunk@1950 83e7a582-4515-4195-8f39-222cbda6297e

+++ merescoharvester/harvester/repository.py
@@ -88,6 +88,9 @@ class HarvestAction(Action):
 
 class DeleteIdsAction(Action):
     def do(self):
+        if self._repository.shopClosed():
+            return False, 'Not deleting outside timeslots.', False
+
         d = DeleteIds(self._repository, self._stateDir, self._logDir, generalHarvestLog=self._generalHarvestLog)
         d.delete()
         return True, 'Deleted', False
